undo: add support for tag to identify an undo record.
This make it possible for a user of the undo circuitry to avoid some
recording as undo data with the same tag won't be recorded. It is
used by the history to avoid too many recording when dragging a curve and
in masks for avoid recording all intermediate results when using the
mouse wheel on forms for example.